+++ contracts/contracts/interfaces/IBeaconProofs.sol
@@ -6,13 +6,13 @@ interface IBeaconProofs {
         bytes32 beaconBlockRoot,
         bytes32 pubKeyHash,
         bytes calldata validatorPubKeyProof,
-        uint64 validatorIndex,
+        uint40 validatorIndex,
         address withdrawalAddress
     ) external view;
 
     function verifyValidatorWithdrawable(
         bytes32 beaconBlockRoot,
-        uint64 validatorIndex,
+        uint40 validatorIndex,
         bytes32 pubKeyHash,
         uint64 withdrawableEpoch,
         bytes calldata withdrawableEpochProof,
@@ -21,7 +21,7 @@ interface IBeaconProofs {
 
     function verifyValidatorWithdrawable(
         bytes32 beaconBlockRoot,
-        uint64 validatorIndex,
+        uint40 validatorIndex,
         uint64 withdrawableEpoch,
         bytes calldata withdrawableEpochProof
     ) external view;
@@ -36,7 +36,7 @@ interface IBeaconProofs {
         bytes32 balancesContainerRoot,
         bytes32 validatorBalanceLeaf,
         bytes calldata balanceProof,
-        uint64 validatorIndex
+        uint40 validatorIndex
     ) external view returns (uint256 validatorBalance);
 
     function verifyFirstPendingDeposit(
